Rendering module organization. Removing redundant commented code. Got rid of warnings.

+++ Modules/FileSystem/Source/FileSystem.cpp
@@ -7,7 +7,7 @@ FileSystem::FileSystem()
 {
 	char* BasePath = SDL_GetBasePath();
 
-	const int Position = FindOccurenceFromString(BasePath, GetPathSeparator(), 4, true);
+	const int Position = static_cast<int>(FindOccurenceFromString(BasePath, GetPathSeparator(), 4, true));
 	char* NewRootDir = new char[Position - 4];
 	memcpy(NewRootDir, BasePath, Position);
